fix: Copilotレビュー指摘事項の修正
- アイテムID生成をレシートID全体を使用する方式に変更しました（36文字→45文字）
  - 任意の位置で切断する問題を解消し、レシートとの関連性を明確に保持します
- bun_repository_test.goのID生成ロジックを修正しました
  - string(rune())の誤用を修正し、fmt.Sprintfで正しく文字列を生成します
- household_usecase.goにオーバーフロー対策を実装しました
  - Price * Quantityの計算時にint64にキャストして安全に計算します
- 大きな値でのオーバーフロー対策を検証するテストを追加しました
- アイテムID長の変更に伴うテストケースを更新しました

+++ internal/modules/household/usecase/household_usecase.go
@@ -59,7 +59,8 @@ func (uc *HouseholdUseCase) GetCategorySummary(ctx context.Context) ([]CategoryS
 				}
 			}
 			summaryMap[category].Count++
-			summaryMap[category].Total += item.Price * item.Quantity
+			// オーバーフロー対策：int64にキャストしてから計算
+			summaryMap[category].Total += int(int64(item.Price) * int64(item.Quantity))
 		}
 	}
